refactor(terminal): extract layout tree helpers to reduce complexity
Extract pure layout-tree operations and keyboard navigation helpers
from TerminalWorkspace.svelte into terminal/layoutTree.ts. This brings
the file under 1000 lines and reduces function complexity below the
max-15 threshold.
