+++ wails-ui/workset/frontend/src/lib/components/TerminalWorkspace.svelte
@@ -9,11 +9,6 @@
 		persistWorkspaceTerminalLayout,
 	} from '../api';
 	import { generateTerminalName } from '../names';
-	import type {
-		TerminalLayout as TerminalLayoutType,
-		TerminalLayoutNode as TerminalLayoutNodeType,
-		TerminalLayoutTab as TerminalLayoutTabType,
-	} from '../types';
 	import {
 		closeTerminal,
 		decreaseFontSize,
@@ -25,6 +20,26 @@
 		resolveTerminalKeybindings,
 		type ResolvedTerminalKeybindings,
 	} from '../terminal/terminalKeybindings';
+	import {
+		type LayoutNode,
+		type TerminalLayout,
+		newId,
+		normalizeLayout,
+		collectTabs,
+		collectPaneIds,
+		findPane,
+		updatePane,
+		splitPane,
+		removePane,
+		updateSplitRatio,
+		moveTab,
+		ensureFocusedPane,
+		applyTabFixes,
+		buildTab,
+		buildPane,
+		buildPanePositions,
+		findAdjacentPane,
+	} from '../terminal/layoutTree';
 
 	interface Props {
 		workspaceId: string;
@@ -34,29 +49,6 @@
 
 	const { workspaceId, workspaceName, active = true }: Props = $props();
 
-	type TerminalTab = TerminalLayoutTabType;
-	type PaneNode = Omit<TerminalLayoutNodeType, 'kind' | 'tabs' | 'activeTabId'> & {
-		kind: 'pane';
-		tabs: TerminalTab[];
-		activeTabId: string;
-		workspaceId?: string;
-		workspaceName?: string;
-	};
-	type SplitNode = Omit<
-		TerminalLayoutNodeType,
-		'kind' | 'first' | 'second' | 'direction' | 'ratio'
-	> & {
-		kind: 'split';
-		first: LayoutNode;
-		second: LayoutNode;
-		direction: 'row' | 'column';
-		ratio: number;
-	};
-	type LayoutNode = PaneNode | SplitNode;
-	type TerminalLayout = Omit<TerminalLayoutType, 'root'> & {
-		root: LayoutNode;
-		focusedPaneId?: string;
-	};
 	const LAYOUT_VERSION = 1;
 	const MIGRATION_VERSION = 1;
 	const SAVE_DEBOUNCE_MS = 300;
@@ -89,91 +81,6 @@
 		} catch {
 			// Ignore storage failures.
 		}
-	};
-
-	const coerceId = (value: unknown): string => {
-		if (typeof value === 'string' && value.trim()) return value;
-		return newId();
-	};
-
-	const normalizeTab = (tab: TerminalTab | undefined | null): TerminalTab | null => {
-		if (!tab) return null;
-		if (typeof tab.id !== 'string' || typeof tab.terminalId !== 'string') return null;
-		const title =
-			typeof tab.title === 'string' && tab.title.trim().length > 0 ? tab.title : 'Terminal';
-		return { id: tab.id, terminalId: tab.terminalId, title };
-	};
-
-	const normalizeNode = (node: TerminalLayoutNodeType | null | undefined): LayoutNode | null => {
-		if (!node || typeof node !== 'object') return null;
-		if (node.kind === 'pane') {
-			const tabs = Array.isArray(node.tabs)
-				? node.tabs.map(normalizeTab).filter((tab): tab is TerminalTab => tab !== null)
-				: [];
-			if (tabs.length === 0) return null;
-			const activeTabId =
-				typeof node.activeTabId === 'string' && tabs.some((tab) => tab.id === node.activeTabId)
-					? node.activeTabId
-					: tabs[0].id;
-			const paneResult: PaneNode = {
-				id: coerceId(node.id),
-				kind: 'pane',
-				tabs,
-				activeTabId,
-			};
-			if (typeof node.workspaceId === 'string' && node.workspaceId) {
-				paneResult.workspaceId = node.workspaceId;
-			}
-			if (typeof node.workspaceName === 'string' && node.workspaceName) {
-				paneResult.workspaceName = node.workspaceName;
-			}
-			return paneResult;
-		}
-		if (node.kind === 'split') {
-			const first = normalizeNode(node.first);
-			const second = normalizeNode(node.second);
-			if (!first && !second) return null;
-			if (!first) return second;
-			if (!second) return first;
-			const direction =
-				node.direction === 'row' || node.direction === 'column' ? node.direction : 'row';
-			const ratio =
-				typeof node.ratio === 'number' &&
-				Number.isFinite(node.ratio) &&
-				node.ratio > 0 &&
-				node.ratio < 1
-					? node.ratio
-					: 0.5;
-			return {
-				id: coerceId(node.id),
-				kind: 'split',
-				direction,
-				ratio,
-				first,
-				second,
-			};
-		}
-		return null;
-	};
-
-	const normalizeLayout = (
-		candidate: TerminalLayoutType | null | undefined,
-	): TerminalLayout | null => {
-		if (!candidate || candidate.version !== LAYOUT_VERSION) return null;
-		const root = normalizeNode(candidate.root);
-		if (!root) return null;
-		return {
-			version: LAYOUT_VERSION,
-			root,
-			focusedPaneId: candidate.focusedPaneId,
-		};
-	};
-
-	const newId = (): string => {
-		if (typeof crypto !== 'undefined' && crypto.randomUUID) {
-			return crypto.randomUUID();
-		}
-		return `term-${Math.random().toString(36).slice(2)}`;
 	};
 
 	const scheduleSaveLayout = (next: TerminalLayout): void => {
@@ -211,7 +118,7 @@
 			const raw = localStorage.getItem(legacyStorageKey(id));
 			if (!raw) return null;
 			const parsed = JSON.parse(raw) as TerminalLayout;
-			return normalizeLayout(parsed);
+			return normalizeLayout(parsed, LAYOUT_VERSION);
 		} catch {
 			return null;
 		}
@@ -221,213 +128,15 @@
 		if (!id) return null;
 		try {
 			const payload = await fetchWorkspaceTerminalLayout(id);
-			return normalizeLayout(payload?.layout);
+			return normalizeLayout(payload?.layout, LAYOUT_VERSION);
 		} catch {
 			return null;
 		}
 	};
 
-	const collectTabs = (node: LayoutNode, tabs: TerminalTab[] = []): TerminalTab[] => {
-		if (node.kind === 'pane') {
-			return tabs.concat(node.tabs);
-		}
-		collectTabs(node.first, tabs);
-		collectTabs(node.second, tabs);
-		return tabs;
-	};
-
-	const collectPaneIds = (node: LayoutNode, ids: string[] = []): string[] => {
-		if (node.kind === 'pane') {
-			ids.push(node.id);
-			return ids;
-		}
-		collectPaneIds(node.first, ids);
-		collectPaneIds(node.second, ids);
-		return ids;
-	};
-
-	const findPane = (node: LayoutNode, paneId: string): PaneNode | null => {
-		if (node.kind === 'pane') {
-			return node.id === paneId ? node : null;
-		}
-		return findPane(node.first, paneId) || findPane(node.second, paneId);
-	};
-
-	const updatePane = (
-		node: LayoutNode,
-		paneId: string,
-		updater: (pane: PaneNode) => PaneNode,
-	): LayoutNode => {
-		if (node.kind === 'pane') {
-			return node.id === paneId ? updater(node) : node;
-		}
-		const first = updatePane(node.first, paneId, updater);
-		const second = updatePane(node.second, paneId, updater);
-		if (first === node.first && second === node.second) {
-			return node;
-		}
-		return { ...node, first, second };
-	};
-
-	const splitPane = (
-		node: LayoutNode,
-		paneId: string,
-		direction: 'row' | 'column',
-		pane: PaneNode,
-	): LayoutNode => {
-		if (node.kind === 'pane') {
-			if (node.id !== paneId) {
-				return node;
-			}
-			return {
-				id: newId(),
-				kind: 'split',
-				direction,
-				ratio: 0.5,
-				first: node,
-				second: pane,
-			};
-		}
-		const first = splitPane(node.first, paneId, direction, pane);
-		const second = splitPane(node.second, paneId, direction, pane);
-		if (first === node.first && second === node.second) {
-			return node;
-		}
-		return { ...node, first, second };
-	};
-
-	const removePane = (node: LayoutNode, paneId: string): LayoutNode | null => {
-		if (node.kind === 'pane') {
-			return node.id === paneId ? null : node;
-		}
-		const first = removePane(node.first, paneId);
-		const second = removePane(node.second, paneId);
-		if (!first && !second) return null;
-		if (!first) return second;
-		if (!second) return first;
-		return { ...node, first, second };
-	};
-
-	const updateSplitRatio = (node: LayoutNode, splitId: string, ratio: number): LayoutNode => {
-		if (node.kind === 'pane') return node;
-		if (node.id === splitId) {
-			return { ...node, ratio };
-		}
-		const first = updateSplitRatio(node.first, splitId, ratio);
-		const second = updateSplitRatio(node.second, splitId, ratio);
-		if (first === node.first && second === node.second) return node;
-		return { ...node, first, second };
-	};
-
-	const moveTab = (
-		node: LayoutNode,
-		sourcePaneId: string,
-		targetPaneId: string,
-		tabId: string,
-		targetIndex: number,
-	): LayoutNode => {
-		// Find the tab in source pane
-		const sourcePane = findPane(node, sourcePaneId);
-		if (!sourcePane) return node;
-		const tab = sourcePane.tabs.find((t) => t.id === tabId);
-		if (!tab) return node;
-
-		// If same pane, just reorder
-		if (sourcePaneId === targetPaneId) {
-			return updatePane(node, sourcePaneId, (pane) => {
-				const tabs = pane.tabs.filter((t) => t.id !== tabId);
-				tabs.splice(targetIndex, 0, tab);
-				return { ...pane, tabs };
-			});
-		}
-
-		// Remove from source
-		let updated = updatePane(node, sourcePaneId, (pane) => ({
-			...pane,
-			tabs: pane.tabs.filter((t) => t.id !== tabId),
-			activeTabId:
-				pane.activeTabId === tabId
-					? (pane.tabs.find((t) => t.id !== tabId)?.id ?? pane.activeTabId)
-					: pane.activeTabId,
-		}));
-
-		// Add to target
-		updated = updatePane(updated, targetPaneId, (pane) => {
-			const tabs = [...pane.tabs];
-			tabs.splice(targetIndex, 0, tab);
-			return { ...pane, tabs, activeTabId: tab.id };
-		});
-
-		return updated;
-	};
-
-	const firstPaneId = (node: LayoutNode): string => {
-		if (node.kind === 'pane') return node.id;
-		return firstPaneId(node.first);
-	};
-
 	const nextTitle = (node: LayoutNode): string => {
 		const count = collectTabs(node).length;
 		return generateTerminalName(workspaceName, count);
-	};
-
-	const buildTab = (terminalId: string, title: string): TerminalTab => ({
-		id: newId(),
-		terminalId,
-		title,
-	});
-
-	const buildPane = (tab: TerminalTab, overrideWorkspaceId?: string, overrideWorkspaceName?: string): PaneNode => {
-		const pane: PaneNode = {
-			id: newId(),
-			kind: 'pane',
-			tabs: [tab],
-			activeTabId: tab.id,
-		};
-		if (overrideWorkspaceId) {
-			pane.workspaceId = overrideWorkspaceId;
-			pane.workspaceName = overrideWorkspaceName;
-		}
-		return pane;
-	};
-
-	const ensureFocusedPane = (next: TerminalLayout): TerminalLayout => {
-		if (!next.focusedPaneId) {
-			return { ...next, focusedPaneId: firstPaneId(next.root) };
-		}
-		if (collectPaneIds(next.root).includes(next.focusedPaneId)) {
-			return next;
-		}
-		return { ...next, focusedPaneId: firstPaneId(next.root) };
-	};
-
-	const applyTabFixes = (
-		node: LayoutNode,
-		fixes: Map<string, { terminalId?: string; drop?: boolean }>,
-	): LayoutNode | null => {
-		if (node.kind === 'pane') {
-			const tabs = node.tabs
-				.map((tab) => {
-					const fix = fixes.get(tab.id);
-					if (fix?.drop) return null;
-					if (fix?.terminalId) {
-						return { ...tab, terminalId: fix.terminalId };
-					}
-					return tab;
-				})
-				.filter((tab): tab is TerminalTab => tab !== null);
-			if (tabs.length === 0) return null;
-			const activeTabId = tabs.some((tab) => tab.id === node.activeTabId)
-				? node.activeTabId
-				: tabs[0].id;
-			return { ...node, tabs, activeTabId };
-		}
-		const first = applyTabFixes(node.first, fixes);
-		const second = applyTabFixes(node.second, fixes);
-		if (!first && !second) return null;
-		if (!first) return second;
-		if (!second) return first;
-		return { ...node, first, second };
 	};
 
 	const migrateLayoutOnce = async (
@@ -511,53 +220,63 @@
 
 	let initToken = 0;
 
+	type StaleGuard = () => boolean;
+
+	const initFromStored = async (
+		wsId: string,
+		stored: TerminalLayout,
+		isStale: StaleGuard,
+	): Promise<void> => {
+		const migrated = await migrateLayoutOnce(wsId, stored);
+		if (isStale()) return;
+		setLayout(migrated.layout);
+		if (migrated.changed) {
+			void persistWorkspaceTerminalLayout(wsId, migrated.layout).catch(() => {});
+		}
+	};
+
+	const initFromLegacy = async (
+		wsId: string,
+		legacy: TerminalLayout,
+		isStale: StaleGuard,
+	): Promise<void> => {
+		const migrated = await migrateLayoutOnce(wsId, legacy);
+		if (isStale()) return;
+		setLayout(migrated.layout);
+		void persistWorkspaceTerminalLayout(wsId, migrated.layout)
+			.then(() => clearLegacyLayout(wsId))
+			.catch(() => {});
+	};
+
+	const initFresh = async (wsId: string, wsName: string, isStale: StaleGuard): Promise<void> => {
+		const created = await createWorkspaceTerminal(wsId);
+		if (isStale()) return;
+		const tab = buildTab(created.terminalId, generateTerminalName(wsName, 0));
+		const pane = buildPane(tab);
+		updateLayout({ version: LAYOUT_VERSION, root: pane, focusedPaneId: pane.id });
+	};
+
 	const initWorkspace = async (): Promise<void> => {
 		if (!workspaceId) return;
 		const token = (initToken += 1);
-		const targetWorkspaceId = workspaceId;
-		const targetWorkspaceName = workspaceName;
+		const targetWsId = workspaceId;
+		const targetWsName = workspaceName;
+		const isStale: StaleGuard = () => token !== initToken || workspaceId !== targetWsId;
 		loading = true;
 		initError = '';
 		layout = null;
 		try {
-			const stored = await loadLayout(targetWorkspaceId);
-			if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-			if (stored) {
-				const migrated = await migrateLayoutOnce(targetWorkspaceId, stored);
-				if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-				if (migrated.changed) {
-					setLayout(migrated.layout);
-					void persistWorkspaceTerminalLayout(targetWorkspaceId, migrated.layout).catch(() => {});
-				} else {
-					setLayout(migrated.layout);
-				}
-				return;
-			}
-			const legacy = loadLegacyLayout(targetWorkspaceId);
-			if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-			if (legacy) {
-				const migrated = await migrateLayoutOnce(targetWorkspaceId, legacy);
-				if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-				setLayout(migrated.layout);
-				void persistWorkspaceTerminalLayout(targetWorkspaceId, migrated.layout)
-					.then(() => {
-						clearLegacyLayout(targetWorkspaceId);
-					})
-					.catch(() => {});
-				return;
-			}
-			const created = await createWorkspaceTerminal(targetWorkspaceId);
-			if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-			const tab = buildTab(created.terminalId, generateTerminalName(targetWorkspaceName, 0));
-			const pane = buildPane(tab);
-			updateLayout({ version: LAYOUT_VERSION, root: pane, focusedPaneId: pane.id });
+			const stored = await loadLayout(targetWsId);
+			if (isStale()) return;
+			if (stored) return await initFromStored(targetWsId, stored, isStale);
+			const legacy = loadLegacyLayout(targetWsId);
+			if (isStale()) return;
+			if (legacy) return await initFromLegacy(targetWsId, legacy, isStale);
+			await initFresh(targetWsId, targetWsName, isStale);
 		} catch (error) {
-			if (token !== initToken || workspaceId !== targetWorkspaceId) return;
-			initError = String(error);
+			if (!isStale()) initError = String(error);
 		} finally {
-			if (token === initToken && workspaceId === targetWorkspaceId) {
-				loading = false;
-			}
+			if (!isStale()) loading = false;
 		}
 	};
 
@@ -660,7 +379,11 @@
 		updateLayout({ ...layout, root: nextRoot });
 	};
 
-	const handleChangePaneWorkspace = async (paneId: string, newWsId: string, newWsName: string): Promise<void> => {
+	const handleChangePaneWorkspace = async (
+		paneId: string,
+		newWsId: string,
+		newWsName: string,
+	): Promise<void> => {
 		if (!layout) return;
 		const pane = findPane(layout.root, paneId);
 		if (!pane) return;
@@ -821,74 +544,78 @@
 		dragState = null;
 	};
 
-	// Keyboard navigation helpers
-	type PanePosition = {
-		id: string;
-		x: number;
-		y: number;
-		w: number;
-		h: number;
+	const handleKeyFocusPane = (event: KeyboardEvent, action: string): void => {
+		if (!layout?.focusedPaneId) return;
+		const direction = action.replace('terminal.focus_pane_', '') as
+			| 'up'
+			| 'down'
+			| 'left'
+			| 'right';
+		event.preventDefault();
+		const positions = buildPanePositions(layout.root);
+		const nextPaneId = findAdjacentPane(layout.focusedPaneId, direction, positions);
+		if (nextPaneId) handleFocusPane(nextPaneId);
 	};
 
-	const buildPanePositions = (
-		node: LayoutNode,
-		x = 0,
-		y = 0,
-		w = 1,
-		h = 1,
-		positions: PanePosition[] = [],
-	): PanePosition[] => {
-		if (node.kind === 'pane') {
-			positions.push({ id: node.id, x, y, w, h });
-			return positions;
-		}
-		const { direction, ratio, first, second } = node;
-		if (direction === 'row') {
-			buildPanePositions(first, x, y, w * ratio, h, positions);
-			buildPanePositions(second, x + w * ratio, y, w * (1 - ratio), h, positions);
-		} else {
-			buildPanePositions(first, x, y, w, h * ratio, positions);
-			buildPanePositions(second, x, y + h * ratio, w, h * (1 - ratio), positions);
-		}
-		return positions;
+	const handleKeyCycleTab = (event: KeyboardEvent, action: string): void => {
+		if (!layout?.focusedPaneId) return;
+		const pane = findPane(layout.root, layout.focusedPaneId);
+		if (!pane || pane.tabs.length <= 1) return;
+		event.preventDefault();
+		const currentIndex = pane.tabs.findIndex((t) => t.id === pane.activeTabId);
+		const delta = action === 'terminal.prev_tab' ? -1 : 1;
+		const nextIndex = (currentIndex + delta + pane.tabs.length) % pane.tabs.length;
+		handleSelectTab(layout.focusedPaneId, pane.tabs[nextIndex].id);
 	};
 
-	const findAdjacentPane = (
-		currentId: string,
-		direction: 'up' | 'down' | 'left' | 'right',
-		positions: PanePosition[],
-	): string | null => {
-		const current = positions.find((p) => p.id === currentId);
-		if (!current) return null;
-
-		const cx = current.x + current.w / 2;
-		const cy = current.y + current.h / 2;
-
-		let candidates = positions.filter((p) => p.id !== currentId);
-
-		// Filter by direction
-		if (direction === 'left') {
-			candidates = candidates.filter((p) => p.x + p.w <= current.x + 0.01);
-		} else if (direction === 'right') {
-			candidates = candidates.filter((p) => p.x >= current.x + current.w - 0.01);
-		} else if (direction === 'up') {
-			candidates = candidates.filter((p) => p.y + p.h <= current.y + 0.01);
-		} else {
-			candidates = candidates.filter((p) => p.y >= current.y + current.h - 0.01);
+	const handleKeyFocusTab = (event: KeyboardEvent, action: string): void => {
+		if (!layout?.focusedPaneId) return;
+		const index = Number.parseInt(action.replace('terminal.focus_tab_', ''), 10);
+		if (!Number.isFinite(index) || index < 1 || index > 9) return;
+		const pane = findPane(layout.root, layout.focusedPaneId);
+		if (!pane) return;
+		const tabIndex = index - 1;
+		if (tabIndex < pane.tabs.length) {
+			event.preventDefault();
+			handleSelectTab(layout.focusedPaneId, pane.tabs[tabIndex].id);
 		}
+	};
 
-		if (candidates.length === 0) return null;
+	const handleKeyCloseTab = (event: KeyboardEvent): void => {
+		if (!layout?.focusedPaneId) return;
+		const pane = findPane(layout.root, layout.focusedPaneId);
+		if (pane?.activeTabId) {
+			event.preventDefault();
+			handleCloseTab(layout.focusedPaneId, pane.activeTabId);
+		}
+	};
 
-		// Find closest by center distance
-		const axis = direction === 'left' || direction === 'right' ? 'y' : 'x';
-		const center = axis === 'x' ? cx : cy;
-		candidates.sort((a, b) => {
-			const aCtr = axis === 'x' ? a.x + a.w / 2 : a.y + a.h / 2;
-			const bCtr = axis === 'x' ? b.x + b.w / 2 : b.y + b.h / 2;
-			return Math.abs(aCtr - center) - Math.abs(bCtr - center);
-		});
+	const withFocusedPane = (event: KeyboardEvent, fn: (paneId: string) => void): void => {
+		if (!layout?.focusedPaneId) return;
+		event.preventDefault();
+		fn(layout.focusedPaneId);
+	};
 
-		return candidates[0].id;
+	const keyActionHandlers: Record<string, (event: KeyboardEvent) => void> = {
+		'terminal.next_tab': (e) => handleKeyCycleTab(e, 'terminal.next_tab'),
+		'terminal.prev_tab': (e) => handleKeyCycleTab(e, 'terminal.prev_tab'),
+		'terminal.close_tab': handleKeyCloseTab,
+		'terminal.new_tab': (e) => withFocusedPane(e, (id) => void handleAddTab(id)),
+		'terminal.split_vertical': (e) => withFocusedPane(e, (id) => void handleSplitPane(id, 'row')),
+		'terminal.split_horizontal': (e) =>
+			withFocusedPane(e, (id) => void handleSplitPane(id, 'column')),
+		'terminal.font_increase': (e) => {
+			e.preventDefault();
+			increaseFontSize();
+		},
+		'terminal.font_decrease': (e) => {
+			e.preventDefault();
+			decreaseFontSize();
+		},
+		'terminal.font_reset': (e) => {
+			e.preventDefault();
+			resetFontSize();
+		},
 	};
 
 	const handleWorkspaceKeydown = (event: KeyboardEvent): void => {
@@ -896,95 +623,9 @@
 		const action = matchTerminalKeybinding(event, resolvedKeybindings);
 		if (!action) return;
 
-		switch (action) {
-			case 'terminal.focus_pane_up':
-			case 'terminal.focus_pane_down':
-			case 'terminal.focus_pane_left':
-			case 'terminal.focus_pane_right': {
-				if (!layout.focusedPaneId) return;
-				const direction = action.replace('terminal.focus_pane_', '') as
-					| 'up'
-					| 'down'
-					| 'left'
-					| 'right';
-				event.preventDefault();
-				const positions = buildPanePositions(layout.root);
-				const nextPaneId = findAdjacentPane(layout.focusedPaneId, direction, positions);
-				if (nextPaneId) {
-					handleFocusPane(nextPaneId);
-				}
-				return;
-			}
-			case 'terminal.next_tab':
-			case 'terminal.prev_tab': {
-				if (!layout.focusedPaneId) return;
-				const pane = findPane(layout.root, layout.focusedPaneId);
-				if (!pane || pane.tabs.length <= 1) return;
-				event.preventDefault();
-				const currentIndex = pane.tabs.findIndex((t) => t.id === pane.activeTabId);
-				const delta = action === 'terminal.prev_tab' ? -1 : 1;
-				const nextIndex = (currentIndex + delta + pane.tabs.length) % pane.tabs.length;
-				handleSelectTab(layout.focusedPaneId, pane.tabs[nextIndex].id);
-				return;
-			}
-			case 'terminal.close_tab': {
-				if (!layout.focusedPaneId) return;
-				const pane = findPane(layout.root, layout.focusedPaneId);
-				if (pane && pane.activeTabId) {
-					event.preventDefault();
-					handleCloseTab(layout.focusedPaneId, pane.activeTabId);
-				}
-				return;
-			}
-			case 'terminal.new_tab': {
-				if (!layout.focusedPaneId) return;
-				event.preventDefault();
-				void handleAddTab(layout.focusedPaneId);
-				return;
-			}
-			case 'terminal.split_vertical': {
-				if (!layout.focusedPaneId) return;
-				event.preventDefault();
-				void handleSplitPane(layout.focusedPaneId, 'row');
-				return;
-			}
-			case 'terminal.split_horizontal': {
-				if (!layout.focusedPaneId) return;
-				event.preventDefault();
-				void handleSplitPane(layout.focusedPaneId, 'column');
-				return;
-			}
-			case 'terminal.font_increase': {
-				event.preventDefault();
-				increaseFontSize();
-				return;
-			}
-			case 'terminal.font_decrease': {
-				event.preventDefault();
-				decreaseFontSize();
-				return;
-			}
-			case 'terminal.font_reset': {
-				event.preventDefault();
-				resetFontSize();
-				return;
-			}
-			default: {
-				if (action.startsWith('terminal.focus_tab_')) {
-					if (!layout.focusedPaneId) return;
-					const index = Number.parseInt(action.replace('terminal.focus_tab_', ''), 10);
-					if (!Number.isFinite(index) || index < 1 || index > 9) return;
-					const pane = findPane(layout.root, layout.focusedPaneId);
-					if (!pane) return;
-					const tabIndex = index - 1;
-					if (tabIndex < pane.tabs.length) {
-						event.preventDefault();
-						handleSelectTab(layout.focusedPaneId, pane.tabs[tabIndex].id);
-					}
-				}
-				return;
-			}
-		}
+		if (action.startsWith('terminal.focus_pane_')) return handleKeyFocusPane(event, action);
+		if (action.startsWith('terminal.focus_tab_')) return handleKeyFocusTab(event, action);
+		keyActionHandlers[action]?.(event);
 	};
 
 	onDestroy(() => {
